Change the reward system
Intercept the access points which include: read, write and mmap accesses
to a file. Keep a track of how many times was a page accessed after
it was evicted from cache. Keep a track of how many time did a page
move back to the active list from the inactive list. Use these
heuristics to decide whether to kick out a page or not.

This is a very basic version. This can be made extremely fine tuned.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>